+++ src/components/feature/detail/BootcampSchedule.tsx
@@ -4,7 +4,7 @@ import {
   Repeat,
   DollarSign,
 } from "lucide-react";
-import type { BootcampDetail } from "@/types/Bootcamp";
+import type { BootcampDetail } from "@/types/response";
 import DetailSectionCard from "./DetailSectionCard";
 
 type BootcampScheduleProps = Pick<
