--- src/types/Bootcamp.ts
@@ -1,42 +1,0 @@
-import type { Review } from "./response";
-
-export interface Bootcamp {
-  bootcampId: number;
-  bootcampName: string;
-  bootcampRegion: string;
-  bootcampCost: boolean;
-  bootcampLink: string;
-  bootcampCategory: string;
-  bootcampDegree: number;
-  bootcampCapacity: number;
-  bootcampStartDate: string;
-  bootcampEndDate: string;
-  courseAverageRating: number;
-  courseReviewCount: number;
-  trainingCenterId: number;
-  trainingCenterName: string;
-  trainingCenterPhoneNumber: string;
-  trainingCenterEmail: string;
-  trainingCenterAddress: string;
-  trainingCenterUrl: string;
-}
-
-export interface BootcampDetail extends Bootcamp {
-  reviews: Review[];
-}
-
-export interface ReviewResponse {
-  content: Review[];
-  pageNumber: number;
-  pageSize: number;
-  totalPages: number;
-  totalElements: number;
-  last: boolean;
-  updatedAt: string;
-}
-
-
-export interface CourseSuggestion {
-  bootcampId: number;
-  bootcampName: string;
-}
