--- src/types/FilterButtons.ts
@@ -1,5 +1,0 @@
-export interface FilterButtonsProps {
-  selectedFilters: Record<string, string>;
-  setSelectedFilters: React.Dispatch<React.SetStateAction<Record<string, string>>>;
-  categoryOptions: string[];
-}
