--- src/types/certification.ts
@@ -1,10 +1,0 @@
-export interface Certification {
-  userId: number;
-  certificationId: number;
-  courseName: string;
-  categoryName: string;
-  trainingCenterName: string;
-  fileUrl: string;
-  userName: string;
-  status: string;
-}
