+++ src/types/response.ts
@@ -1,4 +1,4 @@
-import type { Bootcamp } from "./Bootcamp";
+import { create } from "zustand";
 
 export interface UserInfo {
   email: string;
@@ -96,3 +96,92 @@ export interface ReviewBootcamp {
   categoryName: string;
   trainingProgramId: string;
 }
+
+export interface Bootcamp {
+  bootcampId: number;
+  bootcampName: string;
+  bootcampRegion: string;
+  bootcampCost: boolean;
+  bootcampLink: string;
+  bootcampCategory: string;
+  bootcampDegree: number;
+  bootcampCapacity: number;
+  bootcampStartDate: string;
+  bootcampEndDate: string;
+  courseAverageRating: number;
+  courseReviewCount: number;
+  trainingCenterId: number;
+  trainingCenterName: string;
+  trainingCenterPhoneNumber: string;
+  trainingCenterEmail: string;
+  trainingCenterAddress: string;
+  trainingCenterUrl: string;
+}
+
+export interface BootcampDetail extends Bootcamp {
+  reviews: Review[];
+}
+
+export interface ReviewResponse {
+  content: Review[];
+  pageNumber: number;
+  pageSize: number;
+  totalPages: number;
+  totalElements: number;
+  last: boolean;
+  updatedAt: string;
+}
+
+export interface CourseSuggestion {
+  bootcampId: number;
+  bootcampName: string;
+}
+
+export interface Certification {
+  userId: number;
+  certificationId: number;
+  courseName: string;
+  categoryName: string;
+  trainingCenterName: string;
+  fileUrl: string;
+  userName: string;
+  status: string;
+}
+
+export interface NotificationItem { 
+  notificationId: number;
+  message: string;
+  createdAt: string;
+  type: string;
+  url: string;
+  checked: boolean;
+}
+
+interface NotificationState {
+  notifications: NotificationItem[];
+  setNotifications: (list: NotificationItem[]) => void;
+  markAsReadById: (id: number) => void;
+  addNotification: (item: NotificationItem) => void;
+  markAllAsReadBefore: (time: string) => void;
+}
+
+export const useNotificationStore = create<NotificationState>((set) => ({
+  notifications: [],
+  setNotifications: (list) => set({ notifications: list }),
+  markAsReadById: (id) =>
+    set((s) => ({
+      notifications: s.notifications.map((n) =>
+        n.notificationId === id ? { ...n, checked: true } : n
+      ),
+    })),
+  addNotification: (item) =>
+    set((s) => ({
+      notifications: [item, ...s.notifications],
+    })),
+  markAllAsReadBefore: (time) =>
+    set((s) => ({
+      notifications: s.notifications.map((n) =>
+        n.createdAt <= time ? { ...n, checked: true } : n
+      ),
+    })),
+}));
